got all associates to show up
