+++ index.html
@@ -29,9 +29,9 @@
                     <div class="filter filters__officer"></div>
                 </div>
 
+                <dialog class="knownAssDialogBox"></dialog>
                 <article class="noteFormContainer"></article>
                 <article class="criminalsContainer"></article>
-                <dialog class="knownAssDialogBox"></dialog>
             </div>
             <div class="pushRight">
                 <aside class="noteListContainer">
