+++ scripts/criminals/criminalList.js
@@ -137,14 +137,3 @@ eventHub.addEventListener("officerChosen", event => {
     }
 })
 
-eventHub.addEventListener("click", event => {
-    if (event.target.id.startsWith("associates--")) {
-        const secondHalfOfId = event.target.id.split("--")
-        const customEvent = new CustomEvent("associateAlibiBtnClicked", {
-            detail: {
-                associateId: secondHalfOfId
-            }
-        })
-    eventHub.dispatchEvent(customEvent)    
-    }
-})
